refactor Partner DTOs
